Added a small error detector to paramSweep
If pk.temp == 0 then the output values are NaN. ParamSweep completes fine but then summarizeSweep crashes. Better way of handling this would be to drop values of pK which are not compatable with so few cells, but for now this at least let's it fail gracefully. This bug only really occurs during testing (if less than 2000 cells are used)... but I can imagine quite a few users might try running initially on dummy datasets.